Remove unnecessary GERLSHMUD files
As part of the split from GERLSHMUD to EGRE + EGRE_MUD + EGRE_MUD_1 I'm
removing all files that belong in the other two projects.

The protocol describes an set of rules which is not related to a MUD
engine.

The include files were all specific to a particular MUD

catalog_messages is an old way to figure out what the implicit
"protocol" was. "Protocol" is the event handling rules that are spread
out between different rules processes. e.g. if different processes need
to cooperate to achieve a goal, then the "protocol" is the description
of that communication as found in the code.

Word of warning: coding the logic for events across multiple independent
rules processes is a nightmare. I don't know yet of a situation where
this would be necessary or advisable. I'm doing because it's challenging
and clever. This is a hobby project, so I can be clever; I wouldn't
inflict this on co-workers who were trying to further business goals
with money on the line.