Don't subscribe the egremud_conn to attempts
egremud_conn represents the player's connection state (login, password,
dead, alive) and isn't a graph object. It should propagate messages from
the web socket into the graph with an {attempt, ...} but it should not
try to subscribe to the result (succcess or failure).

As a reminder, there are three processes for each player connection:
- web socket / test socket -> egremud_conn state machine -> graph object

We also left in a convenient ct:pal call for getting the PIDs of those
three processes above.