Connect player2 to a socket by its login
In order to test functionality between players it's necessary to have
both players login. 'player' was logged in automatically, so we left
that, but we added a credentials check clause so that <<"player2">> can
be specified as a login.

This goes along with a change to the test socket for specifying what the
registered ID of the process should be, so that sending and fetching
messages can be targeted at a specific player's connection.