Allow starting multiple distinct test sockets
Now you can start a non-default test socket and give it an ID. The ID
has nothing to do with the player that is attached to the socket, but
it's convenient if you use the same socket ID and graph object ID, e.g.
player2, in both cases. The player attached to the socket is based on
the login used.

Also add the PID of the socket when logging received messages so you can
tell which player received what.