Register the websocket as a json logger
When I've run the game in the past it was nice to have the log messages
come into the log web page in real time. This adds a running websocket
as a log sink.

Since EGRE doesn't know anything about web servers, or web sockets,
being only a graph rules engine, EGRE_MUD needs to register the
websocket as a log listener.